Rename filename of AlifyClientService
